Creating GeomTools class
